requirements.txt
alembic==1.13.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
billiard==4.2.1
black==24.8.0
celery==5.5.1
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.1
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
git-filter-repo==2.45.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
iniconfig==2.1.0
Jinja2==3.1.4
kombu==5.5.3
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.6
pluggy==1.6.0
prompt_toolkit==3.0.51
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.9
PyYAML==6.0.2
redis==5.2.1
rich==13.8.1
ruff==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
time-machine==2.16.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
